OSPL-3310 Warning free on windows 64 bit (requires fix OSPL-3310 to SAC codegen aso - see mainline).

+++ touchstone/c/metrics.c
@@ -72,7 +72,7 @@ metrics_report (
         report->maximum = 0;
         report->deviation = 0;
     } else {
-        rng = (cnt*prc)/100;
+        rng = (int) ((cnt*prc)/100);
         min = lst[0];
         max = lst[rng-1];
         sum = 0.0;
